==> 01 Exploratory data analysis.ipynb <==
# Titanic Kaggle Competition

This notebook contains a little bit of exploratory data analysis of the Titanic-dataset. The goal is to find out if there are features in the dataset that have strong correlation with survival of a passenger.

https://www.kaggle.com/c/titanic

# Feature engineering and data exploration

import pandas as pd
import seaborn as sns

dataset = pd.read_csv('train.csv')
testset = pd.read_csv('test.csv')
dataset = dataset.drop(['PassengerId'], axis=1)
testset = testset.drop(['PassengerId'], axis=1)

dataset.head()

On the Kaggle's instruction video on the Titanic competition, it is said that female passengers were more likely to survive the wreck compared to male passengers. Let's explore this...

dataset.groupby('Sex').mean()

As it can be seen from the table above, at least on the training dataset, almost 75% of the female passengers survived, whereas only about 19% of the male passengers survived. So, 'Sex' feature will probably have strong effect on predicting who will survive.

Let's explore if the tickect class correlates with the survival of a passenger.

dataset.groupby('Pclass').mean()

It seems that passengers of the ticket class 1 have bigger chance of surviving compared to the passengers of the other two classes.

Passengers embarked to Titanic from three locations, Cherbourg (C), Queenstown (Q), and Southampton (S). Let's explore if the port of embarkment correlates with the survival of a passenger.

Let's first check though, if the port of embarkement is known for each passenger. And also find out if there is missing values on other features.

dataset.info()

Features 'Age', 'Cabin' and 'Embarked' contain empty values. Although, the port of embarkement data is missing only for two passengers, so let's choose the most frequent value for them. Since almost 800 passengers lack the 'Cabin' data, let's just drop that column.

testset.info()

The testset also contains one null Fare.

dataset['Age'].describe()

# Since the spread of 'Age' is large, median is a better choice than mean.
dataset['Age'] = dataset['Age'].fillna(dataset['Age'].median())
testset['Age'] = testset['Age'].fillna(testset['Age'].median())
testset['Fare'] = testset['Fare'].fillna(testset['Fare'].median())

dataset['Embarked'].value_counts()

def emb(e):
    good = ['S', 'C', 'Q']

    if e not in good:
        return 'S'
    else:
        return e

dataset['Embarked'] = dataset['Embarked'].apply(emb)
testset['Embarked'] = testset['Embarked'].apply(emb)
dataset = dataset.drop(['Cabin'], axis=1)
testset = testset.drop(['Cabin'], axis=1)

dataset.groupby('Embarked').mean()

As we can see, passengers who embarked from Cherbourg have higher chance of surviving compared to passengers who embarked from the other two ports.

The spread of the 'Fare', ie. ticket price, is also wide. Let's round the 'Fare' to whole number in order to group the fares more easily.

dataset['Fare'].describe()

dataset['RoundFare'] = dataset['Fare'].round()
testset['RoundFare'] = testset['Fare'].round()

dataset.head()

dataset.groupby('RoundFare').mean()

dataset['RoundFare'].value_counts()

counts = dataset['RoundFare'].value_counts()

for fare, count in counts.items():
  print(fare, count)

With a quick glance we can see that for a little bit over half of the passengers (472) the fare was between 7 and 16. There's also 22 passengers with the fare price of under 7. So the first fare price group is 0-16. Second group is fare prices from 17 to 31, this consist of 186 passengers, or about 20%. For the rest of the fare prices, no bigger clusters can be found, so the third group is "the rest", ie. over 31, up to 512.

def age_group(age):
  if age < 17:
    return "group_1"
  elif age < 32:
    return "group_2"
  else:
    return "group_3"

dataset['AgeGroup'] = dataset['Age'].apply(age_group)
testset['AgeGroup'] = testset['Age'].apply(age_group)

dataset.head()

dataset['AgeGroup'].value_counts() / dataset.shape[0]

dataset.groupby('AgeGroup').mean()

As we can see above, the younger the passenger, the bigger the chance of surviving.

Feature 'sibsp' is the amount of siblings / spouses on board, and 'parch' is amount of parents / children on board. Let's combine them as a new feature 'FamilySize'.

dataset['FamilySize'] = dataset['SibSp'] + dataset['Parch']
testset['FamilySize'] = testset['SibSp'] + testset['Parch']

dataset.head()

dataset.groupby('FamilySize').mean()

dataset['FamilySize'].value_counts()

It seems that passengers with small "entourage" with them had the biggest chance of surviving. The bigger the "entourage", the smaller the change of survival. Also over half of the passengers travelled alone, only 30% of them surviving.

Each passenger seem to have some sort of a title included with their names. Let's see if it correlates with survival.

Since over half of the passengers travelled alone, et's create one more new feature 'TravelledAlone'.

def alone(family_size):
  if family_size == 0:
    return "alone"
  else:
    return "not_alone"

dataset['TravelledAlone'] = dataset['FamilySize'].apply(alone)
testset['TravelledAlone'] = testset['FamilySize'].apply(alone)

dataset.head()

dataset.groupby('TravelledAlone').mean()

Travelling alone was more riskier, even though also only half of the passengers not alone, survived.

def title(name):
  return name.split(".")[0].split(",")[1].strip()

dataset['Title'] = dataset['Name'].apply(title)
testset['Title'] = testset['Name'].apply(title)

dataset['Title'].value_counts()

Let's for five title groups: Mr, Miss, Mrs, Master and Other. (in the test dataset, there is also one person with title Dona.)

def title_group(title):
  rest = ["Dr", "Rev", "Mlle", "Major", "Col", "the Countess", "Mme", "Jonkheer", "Sir", "Lady", "Ms", "Don", "Dona", "Capt"]

  if title not in rest:
    return title
  else:
    return "Other"

dataset['TitleGroup'] = dataset['Title'].apply(title_group)
testset['TitleGroup'] = testset['Title'].apply(title_group)

dataset['TitleGroup'].value_counts()

dataset.groupby('TitleGroup').mean()

So once again: women and children are way more likely to survive than men.

dataset.head()

Let's drop the columns / features we don't need anymore. Title can be dropped, because the exact same information is contained in TitleGroup. Ticket has too many different options, so it is impossible to derive any features from it. Also, name is not needed anymore.

dataset = dataset.drop(['Name', 'Title', 'Ticket'], axis=1)
testset = testset.drop(['Name', 'Title', 'Ticket'], axis=1)

dataset.to_csv('dataset.csv', index=None)
testset.to_csv('testset.csv', index=None)
