==> 10 Titanic Ensemble.ipynb <==
# Ensembling

On this notebook, all the predictions from the models presented before, are combined to one dataframe. Then a simple voting is executed, where the majority vote will be the prediction for each testset item.

import os
import pandas as pd

# First try

Combine all the model predictions into a single dataframe.

folder = "csvs/"
predictions = pd.DataFrame()
column_count = 0

for index, filename in enumerate(os.listdir(folder)):
  if filename.endswith('.csv'):
    df = pd.read_csv(folder + filename)

    if "PassengerId" not in predictions:
      predictions = df
    else:
      predictions["col_" + str(index)] = df['Survived']

    column_count = index + 1


predictions.rename(columns={"Survived": "col_0"}, inplace=True)
predictions.head()


# This function gets the "majority vote" of all the model predictions for each row. 
def get_vote(row):
  if row.drop('PassengerId').sum() > column_count / 2:
    return 1
  else:
    return 0

# Get the ensembled prediction for each testset item.
predictions['vote'] = predictions.apply(get_vote, axis=1)

predictions.head()

Separate the final ensembled predictions into a dataframe.

predictions.rename(columns={"vote": "Survived"}, inplace=True)
ensembled_predictions = predictions[['PassengerId', 'Survived']]
ensembled_predictions.head()

ensembled_predictions.to_csv('ensemble_preds_02.csv', index=None)

When taking into account the predictions of each model, we get a an accuracy of 0.784, which is the best accuracy so far.

# Second try

Let's try if we can still improve our accuracy by taking the majority vote of only the three best models: Support Vector Machine, Kernel SVM and Logistic Regression.

svm = pd.read_csv('csvs/titanic_svm_01.csv')
kernel_svm = pd.read_csv('csvs/titanic_kernel_svm_01.csv')
lgreg = pd.read_csv('csvs/titanic_logistic_regression_01.csv')

predictions2 = svm
predictions2.rename(columns={"Survived": "col_0"}, inplace=True)
predictions2['col_1'] = kernel_svm['Survived']
predictions2['col_2'] = lgreg['Survived']

predictions2.head(20)

def get_vote2(row):
  if row.drop('PassengerId').sum() > 2:
    return 1
  else:
    return 0

predictions2['Survived'] = predictions2.apply(get_vote2, axis=1)

predictions2.head()

ensemble2 = predictions2[['PassengerId', 'Survived']]
ensemble2.head()

ensemble2.to_csv('best_ensemble.csv', index=None)

This method of ensembling results an accuracy of only 0.777.
